requirements.txt
discord.py>=2.0.0
Flask>=2.0.0
Flask-CORS>=3.0.0
python-dotenv>=0.19.0
requests>=2.26.0
flask-login>=0.5.0
Werkzeug>=2.0.0
